Fix how MHD speeds are calculated
Treat them as scalars, rather than trying to compare them to v_θ like
the sound speed.
