+++ src/disc_solver/solve/utils.py
@@ -12,7 +12,6 @@ from numpy import (
     any as np_any,
     diff,
     abs as np_abs,
-    sqrt as np_sqrt,
     max as np_max,
 )
 from scipy.interpolate import interp1d
@@ -21,8 +20,7 @@ from scikits.odes.sundials.cvode import StatusEnum
 
 from ..file_format import CONFIG_FIELDS, SOLUTION_INPUT_FIELDS
 from ..utils import (
-    ODEIndex, DiscSolverError, MHD_Wave_Index, mhd_wave_speeds,
-    MAGNETIC_INDEXES, expanded_path
+    ODEIndex, DiscSolverError, expanded_path,
 )
 
 log = logbook.Logger(__name__)
@@ -32,25 +30,6 @@ def error_handler(error_code, module, func, msg, user_data):
     """ drop all CVODE/IDA messages """
     # pylint: disable=unused-argument
     pass
-
-
-def gen_mhd_sonic_point_rootfn(mhd_wave_type):
-    """
-    Finds specified mhd sonic point
-    """
-    mhd_wave_type = MHD_Wave_Index[mhd_wave_type]
-
-    def rootfn(θ, params, out):
-        """
-        root function to find mhd sonic point
-        """
-        # pylint: disable=unused-argument
-        wave_speeds = np_sqrt(mhd_wave_speeds(
-            params[MAGNETIC_INDEXES], params[ODEIndex.ρ], 1
-        ))
-        out[0] = wave_speeds[mhd_wave_type] - params[ODEIndex.v_θ]
-        return 0
-    return rootfn
 
 
 def validate_solution(soln):
